Graceful handling of closed gist windows
If the gist export takes a long time, the student may close the window
before it completes (or fails). We want to handle this gracefully:

* If the export completes and the student has closed the window, we
  display an application error instructing the student to try again.
* If the export fails and the student has closed the window, we don’t
  try to close the window again.

Also used this as an opportunity to move the meat of the gist export
logic out of the `Dashboard` and into the `Workspace`, since the latter
as the view controller should take care of (or delegate to the action
creators) pretty much anything interesting.

Threw in another small fix which is to always open a new tab for the
exported gist, rather than reusing the same tab for successive exports
(which is 100% confusing).
